in process

+++ Magazin/BaseCore.Security/Entities/Role.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace BaseCore.Security.Entities
 {
-    public class Role : IdentityRole<int>
+    public class Role : IdentityRole<long>
     {
     }
 }
